+++ src/main/kotlin/sinhee/kang/tutorial/infra/api/vision/service/VisionServiceImpl.kt
@@ -1,4 +1,4 @@
-package sinhee.kang.tutorial.infra.api.vision
+package sinhee.kang.tutorial.infra.api.vision.service
 
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import okhttp3.*
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Component
 import org.springframework.web.multipart.MultipartFile
 import retrofit2.Retrofit
 import retrofit2.converter.jackson.JacksonConverterFactory
+import sinhee.kang.tutorial.infra.api.vision.KakaoApi
 
 @Component
 class VisionServiceImpl(
@@ -18,7 +19,7 @@ class VisionServiceImpl(
             .baseUrl("https://dapi.kakao.com/")
             .addConverterFactory(JacksonConverterFactory.create(jacksonObjectMapper()))
             .build()
-        .create(KakaoApiInterface::class.java)
+        .create(KakaoApi::class.java)
 
     override fun generateTagFromImage(imageFile: MultipartFile): List<String> {
         val requestFile: RequestBody = RequestBody.create(MediaType.parse("image/*"), imageFile.bytes)
